test change

+++ src/Utility/Utils.java
@@ -16,7 +16,7 @@ public class Utils {
 		BigDecimal SUM, AVERAGE;
 	
            
-		long sum = 0;
+		double sum = 0.0;
 	
 		for(int i=0; i < runtime.length ; i++)
             sum = sum + runtime[i];
